fix: Better coercion messages for change-default
Also: move server/responses to their own files
